# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AiActionModeUpdate": {
        "properties": {
          "mode": {
            "description": "notify | auto",
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode"
        ],
        "title": "AiActionModeUpdate",
        "type": "object"
      },
...
      "AiSettingsUpdate": {
        "properties": {
          "api_key": {
            "default": "",
            "maxLength": 256,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "default": "",
            "maxLength": 256,
            "title": "Base Url",
            "type": "string"
          },
          "model": {
            "default": "",
            "maxLength": 64,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider"
        ],
        "title": "AiSettingsUpdate",
        "type": "object"
      },
      "AiSpendCapUpdate": {
        "properties": {
          "cap": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Cap"
          }
        },
        "required": [
          "cap"
        ],
        "title": "AiSpendCapUpdate",
        "type": "object"
      },
...
      "BinanceKeysUpdate": {
        "properties": {
          "api_key": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "api_secret": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Api Secret",
            "type": "string"
          }
        },
        "required": [
          "api_key",
          "api_secret"
        ],
        "title": "BinanceKeysUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LlmCredentialsUpdate": {
        "description": "One LLM provider's credentials \u2014 both fields optional.",
        "properties": {
          "api_key": {
            "default": "",
            "maxLength": 256,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "default": "",
            "maxLength": 256,
            "title": "Base Url",
            "type": "string"
          }
        },
        "title": "LlmCredentialsUpdate",
        "type": "object"
      },
...
      "ModeUpdate": {
        "properties": {
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/TradingMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "ModeUpdate",
        "type": "object"
      },
...
      "SettingsRead": {
        "properties": {
          "ai_action_mode": {
            "title": "Ai Action Mode",
            "type": "string"
          },
          "ai_base_url": {
            "title": "Ai Base Url",
            "type": "string"
          },
          "ai_key_configured": {
            "title": "Ai Key Configured",
            "type": "boolean"
          },
          "ai_model": {
            "title": "Ai Model",
            "type": "string"
          },
          "ai_provider": {
            "title": "Ai Provider",
            "type": "string"
          },
          "ai_spend_cap": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Ai Spend Cap",
            "type": "string"
          },
          "ai_spend_today": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Ai Spend Today",
            "type": "string"
          },
          "binance_keys_configured": {
            "title": "Binance Keys Configured",
            "type": "boolean"
          },
          "llm_providers_configured": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Llm Providers Configured",
            "type": "object"
          },
          "mode": {
            "$ref": "#/components/schemas/TradingMode"
          },
          "venue_credentials_configured": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Venue Credentials Configured",
            "type": "object"
          }
        },
        "required": [
          "mode",
          "binance_keys_configured",
          "venue_credentials_configured",
          "ai_provider",
          "ai_model",
          "ai_base_url",
          "ai_key_configured",
          "ai_spend_cap",
          "ai_spend_today",
          "ai_action_mode",
          "llm_providers_configured"
        ],
        "title": "SettingsRead",
        "type": "object"
      },
...
      "TradingMode": {
        "description": "Where orders are routed. Sim is the safe default.",
        "enum": [
          "sim",
          "testnet",
          "live"
        ],
        "title": "TradingMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VenueCredentialsUpdate": {
        "description": "A venue's credential fields \u2014 names validated against the catalogue.",
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "minProperties": 1,
            "title": "Fields",
            "type": "object"
          }
        },
        "required": [
          "fields"
        ],
        "title": "VenueCredentialsUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "summary": "Self-hosted automated Binance trading engine.",
    "title": "Capital Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "description": "Current trading mode and whether Binance keys are stored.",
        "operationId": "read_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/ai": {
      "put": {
        "description": "Configure the AI provider. The API key is encrypted at rest.",
        "operationId": "update_ai_settings_api_settings_ai_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Ai Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/ai-action-mode": {
      "put": {
        "description": "Set how AI decisions are applied: `notify` (confirm) or `auto`.",
        "operationId": "update_ai_action_mode_api_settings_ai_action_mode_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiActionModeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Ai Action Mode",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/ai-spend-cap": {
      "put": {
        "description": "Set the daily LLM spend cap \u2014 AI strategies pause once it is reached.",
        "operationId": "update_ai_spend_cap_api_settings_ai_spend_cap_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSpendCapUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Ai Spend Cap",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/binance-keys": {
      "put": {
        "description": "Store the Binance API credentials (encrypted at rest).",
        "operationId": "update_binance_keys_api_settings_binance_keys_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BinanceKeysUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Binance Keys",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/llm-credentials/{provider}": {
      "put": {
        "description": "Store one LLM provider's API key and/or base URL (encrypted at rest).",
        "operationId": "update_llm_credentials_api_settings_llm_credentials__provider__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmCredentialsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Llm Credentials",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/mode": {
      "put": {
        "description": "Switch the trading mode (Sim / Testnet / Live).",
        "operationId": "update_mode_api_settings_mode_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Mode",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/venue-credentials/{venue}": {
      "put": {
        "description": "Store a venue's credentials (encrypted at rest).\n\nThe submitted field names must exactly match the venue's declared\n`credential_fields`, and every value must be non-empty.",
        "operationId": "update_venue_credentials_api_settings_venue_credentials__venue__put",
        "parameters": [
          {
            "in": "path",
            "name": "venue",
            "required": true,
            "schema": {
              "title": "Venue",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueCredentialsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Venue Credentials",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}